+++ Sources/SuperwallKit/Dependencies/DependencyContainer.swift
@@ -142,7 +142,11 @@ final class DependencyContainer {
       storage: storage,
       configManager: configManager,
       webEntitlementRedeemer: webEntitlementRedeemer
-    )
+    ) { [weak self] newAttributes in
+      Task { @MainActor in
+        self?.delegateAdapter.userAttributesDidChange(newAttributes: newAttributes)
+      }
+    }
 
     appSessionManager = AppSessionManager(
       configManager: configManager,
@@ -275,6 +279,7 @@ extension DependencyContainer: ViewControllerFactory {
       deviceHelper: deviceHelper,
       factory: self,
       storage: storage,
+      network: network,
       webView: webView,
       webEntitlementRedeemer: webEntitlementRedeemer,
       cache: cache,