+++ CHANGELOG.md
@@ -2,10 +2,27 @@
 
 The changelog for `SuperwallKit`. Also see the [releases](https://github.com/superwall/Superwall-iOS/releases) on GitHub.
 
-## 4.10.7
+## 4.11.0
+
+### Enhancements
+
+- Adds the ability to override introductory offer eligibility via the paywall editor.
 
 ### Fixes
 
+- Fixes an issue where not all product IDs belonging to `Entitlement`s in `CustomerInfo` were being included.
+
+## 4.10.8
+
+### Enhancements
+
+- Adds support for `Set user attributes` action.
+- Adds new `SuperwallDelegate` method called `userAttributesDidChange` that notifies you when user attributes change from an external source.
+- Adds `firebaseInstallationId` as an `IntegrationAttribute`.
+
+### Fixes
+
+- Fixes a crash caused by a race condition when accessing JSON dictionaries concurrently.
 - Fixes issue returning the `PurchaseResult` from `Superwall.shared.purchase(_:)` when using StoreKit 1 inside a `PurchaseController`.
 - Fixes `handleDeepLink` returning true for non-Superwall URLs when called before configuration completes.